# docs/openapi.yaml
openapi: 3.0.3
info:
  title: gator API
  version: 0.1.0
  description: >
    HTTP surface of gator-server. Every route lives under /api/v1. Live updates are delivered
    over the WebSocket at /api/v1/ws (topics: inbox, task:<id>, job:<id>).
servers:
  - url: /api/v1
paths:
  /healthz:
    get:
      operationId: healthz
      responses:
        "200": { description: ok, content: { application/json: { schema: { $ref: "#/components/schemas/Status" } } } }
  /readyz:
    get:
      operationId: readyz
      responses:
        "200": { description: ready, content: { application/json: { schema: { $ref: "#/components/schemas/Readiness" } } } }
        "503": { description: not ready, content: { application/json: { schema: { $ref: "#/components/schemas/Readiness" } } } }
  /version:
    get:
      operationId: version
      responses:
        "200": { description: ok, content: { application/json: { schema: { $ref: "#/components/schemas/Version" } } } }
  /capabilities:
    get:
      operationId: capabilities
      responses:
        "200": { description: ok, content: { application/json: { schema: { $ref: "#/components/schemas/Capabilities" } } } }
  /auth/me:
    get:
      operationId: me
      responses:
        "200": { description: ok, content: { application/json: { schema: { $ref: "#/components/schemas/Me" } } } }
        "401": { $ref: "#/components/responses/Error" }
  /tokens:
    get:
      operationId: listTokens
      description: Tokens issued to the caller (users) or every runner/plugin token (workspace admins).
      responses:
        "200": { description: ok, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/Token" } } } } }
    post:
      operationId: createToken
      description: Mint a bearer token. Runner and plugin tokens require workspace admin; user tokens are for the caller.
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/NewToken" } } }
      responses:
        "201": { description: created, content: { application/json: { schema: { $ref: "#/components/schemas/IssuedToken" } } } }
        "403": { $ref: "#/components/responses/Error" }
  /tokens/{tokenId}:
    delete:
      operationId: revokeToken
      parameters: [ { name: tokenId, in: path, required: true, schema: { type: string, format: uuid } } ]
      responses:
        "204": { description: revoked }
        "403": { $ref: "#/components/responses/Error" }
  /projects/{projectId}/members:
    put:
      operationId: setMember
      description: Set a user's role in a project. Requires project admin.
      parameters: [ { $ref: "#/components/parameters/projectId" } ]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/Member" } } }
      responses:
        "204": { description: ok }
        "403": { $ref: "#/components/responses/Error" }
  /projects:
    get:
      operationId: listProjects
      responses:
        "200": { description: ok, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/Project" } } } } }
    post:
      operationId: createProject
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/NewProject" } } }
      responses:
        "201": { description: created, content: { application/json: { schema: { $ref: "#/components/schemas/Project" } } } }
        "400": { $ref: "#/components/responses/Error" }
  /projects/{projectId}:
    get:
      operationId: getProject
      parameters: [ { $ref: "#/components/parameters/projectId" } ]
      responses:
        "200": { description: ok, content: { application/json: { schema: { $ref: "#/components/schemas/Project" } } } }
        "404": { $ref: "#/components/responses/Error" }
  /projects/{projectId}/tasks:
    get:
      operationId: listTasks
      parameters: [ { $ref: "#/components/parameters/projectId" } ]
      responses:
        "200": { description: ok, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/Task" } } } } }
    post:
      operationId: createTask
      parameters: [ { $ref: "#/components/parameters/projectId" } ]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/NewTask" } } }
      responses:
        "201": { description: created, content: { application/json: { schema: { $ref: "#/components/schemas/Task" } } } }
        "400": { $ref: "#/components/responses/Error" }
  /inbox:
    get:
      operationId: listInbox
      description: Decisions waiting for a human, most urgent first.
      parameters:
        - name: projectId
          in: query
          required: false
          schema: { type: string, format: uuid }
      responses:
        "200": { description: ok, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/Decision" } } } } }
  /tasks/{taskId}:
    get:
      operationId: getTask
      parameters: [ { $ref: "#/components/parameters/taskId" } ]
      responses:
        "200": { description: ok, content: { application/json: { schema: { $ref: "#/components/schemas/TaskDetail" } } } }
        "404": { $ref: "#/components/responses/Error" }
  /tasks/{taskId}/approve:
    post:
      operationId: approveTask
      parameters: [ { $ref: "#/components/parameters/taskId" } ]
      responses:
        "200": { description: ok, content: { application/json: { schema: { $ref: "#/components/schemas/Task" } } } }
        "409": { $ref: "#/components/responses/Error" }
  /tasks/{taskId}/advance:
    post:
      operationId: advanceTask
      parameters: [ { $ref: "#/components/parameters/taskId" } ]
      requestBody:
        content: { application/json: { schema: { $ref: "#/components/schemas/Reason" } } }
      responses:
        "200": { description: ok, content: { application/json: { schema: { $ref: "#/components/schemas/Task" } } } }
        "409": { $ref: "#/components/responses/Error" }
  /tasks/{taskId}/rollback:
    post:
      operationId: rollbackTask
      parameters: [ { $ref: "#/components/parameters/taskId" } ]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/Rollback" } } }
      responses:
        "200": { description: ok, content: { application/json: { schema: { $ref: "#/components/schemas/Task" } } } }
        "409": { $ref: "#/components/responses/Error" }
  /tasks/{taskId}/handoff:
    post:
      operationId: handoffTask
      parameters: [ { $ref: "#/components/parameters/taskId" } ]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/Handoff" } } }
      responses:
        "200": { description: ok, content: { application/json: { schema: { $ref: "#/components/schemas/Task" } } } }
        "409": { $ref: "#/components/responses/Error" }
  /tasks/{taskId}/checks:
    put:
      operationId: setTaskCheck
      description: Record an automated verdict on the current gate. Used by plugins.
      parameters: [ { $ref: "#/components/parameters/taskId" } ]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/Check" } } }
      responses:
        "200": { description: ok, content: { application/json: { schema: { $ref: "#/components/schemas/Task" } } } }
        "409": { $ref: "#/components/responses/Error" }
  /tasks/{taskId}/transitions:
    get:
      operationId: listTaskTransitions
      parameters: [ { $ref: "#/components/parameters/taskId" } ]
      responses:
        "200": { description: ok, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/Transition" } } } } }
components:
  parameters:
    projectId: { name: projectId, in: path, required: true, schema: { type: string, format: uuid } }
    taskId: { name: taskId, in: path, required: true, schema: { type: string, format: uuid } }
  responses:
    Error:
      description: error
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  schemas:
    Error:
      type: object
      required: [error]
      properties: { error: { type: string }, code: { type: string } }
    Status:
      type: object
      required: [status]
      properties: { status: { type: string } }
    Readiness:
      type: object
      required: [checks]
      properties: { checks: { type: object, additionalProperties: { type: string } } }
    Version:
      type: object
      required: [version, commit, date]
      properties: { version: { type: string }, commit: { type: string }, date: { type: string } }
    Capabilities:
      type: object
      required: [features, protocolVersion]
      properties:
        features: { type: array, items: { type: string } }
        protocolVersion: { type: integer }
    Me:
      type: object
      required: [kind, name, workspaceRole]
      properties:
        kind: { type: string }
        userId: { type: string, format: uuid }
        name: { type: string }
        workspaceRole: { type: string }
        scope: { type: string }
    NewToken:
      type: object
      required: [kind, name]
      properties:
        kind: { type: string, enum: [user, runner, plugin] }
        name: { type: string }
        projectId: { type: string, format: uuid, description: required for plugin tokens }
        ttlHours: { type: integer }
    Token:
      type: object
      required: [id, kind, scope, name, createdAt]
      properties:
        id: { type: string, format: uuid }
        kind: { type: string }
        scope: { type: string }
        name: { type: string }
        projectId: { type: string, format: uuid }
        createdAt: { type: string, format: date-time }
        expiresAt: { type: string, format: date-time }
        lastUsedAt: { type: string, format: date-time }
    IssuedToken:
      allOf:
        - $ref: "#/components/schemas/Token"
        - type: object
          required: [token]
          properties:
            token: { type: string, description: plaintext, shown once }
    Member:
      type: object
      required: [userId, role]
      properties:
        userId: { type: string, format: uuid }
        role: { type: string, enum: [admin, member, viewer] }
    NewProject:
      type: object
      required: [slug, name]
      properties:
        slug: { type: string }
        name: { type: string }
        tags: { type: array, items: { type: string } }
    Project:
      type: object
      required: [id, slug, name, tags, createdAt]
      properties:
        id: { type: string, format: uuid }
        slug: { type: string }
        name: { type: string }
        tags: { type: array, items: { type: string } }
        createdAt: { type: string, format: date-time }
    NewTask:
      type: object
      required: [kind, title]
      properties:
        kind: { type: string, enum: [feature, bug, incident, chore] }
        title: { type: string }
        urgency: { type: integer, minimum: 1, maximum: 4 }
    Task:
      type: object
      required: [id, projectId, kind, title, phase, urgency, requirementsChanged, phaseEnteredAt, createdAt]
      properties:
        id: { type: string, format: uuid }
        projectId: { type: string, format: uuid }
        kind: { type: string }
        title: { type: string }
        phase: { type: string }
        urgency: { type: integer }
        ownerKind: { type: string }
        ownerId: { type: string, format: uuid }
        requirementsChanged: { type: boolean }
        blockedReason: { type: string }
        phaseEnteredAt: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }
        closedAt: { type: string, format: date-time }
    TaskDetail:
      allOf:
        - $ref: "#/components/schemas/Task"
        - type: object
          required: [gate, phases]
          properties:
            gate: { $ref: "#/components/schemas/Gate" }
            phases: { type: array, items: { type: string }, description: active phases of this task's template in order }
    Gate:
      type: object
      required: [phase, kind, checks, humanApproved]
      properties:
        phase: { type: string }
        kind: { type: string, enum: [human, auto, both] }
        checks: { type: array, items: { $ref: "#/components/schemas/Check" } }
        humanApproved: { type: boolean }
        blockedReason: { type: string }
        blockedBy: { type: string }
    Check:
      type: object
      required: [name, source, status]
      properties:
        name: { type: string }
        source: { type: string }
        status: { type: string, enum: [pass, fail, pending] }
        detail: { type: string }
    Decision:
      type: object
      required: [task, waitingSince, reason]
      properties:
        task: { $ref: "#/components/schemas/Task" }
        waitingSince: { type: string, format: date-time }
        reason: { type: string, enum: [approval, blocked, requirements_changed, idea] }
    Reason:
      type: object
      properties: { reason: { type: string } }
    Rollback:
      type: object
      required: [to, reason]
      properties: { to: { type: string }, reason: { type: string } }
    Handoff:
      type: object
      required: [toKind, toId]
      properties:
        toKind: { type: string, enum: [user, runner] }
        toId: { type: string, format: uuid }
        reason: { type: string }
    Transition:
      type: object
      required: [id, toPhase, kind, actorKind, createdAt]
      properties:
        id: { type: integer, format: int64 }
        fromPhase: { type: string }
        toPhase: { type: string }
        kind: { type: string }
        actorKind: { type: string }
        actorId: { type: string, format: uuid }
        reason: { type: string }
        createdAt: { type: string, format: date-time }